Update from `generate-script` to `ndm generate`.
`generate-script` does not exists in this container's PATH. There is no file named generate-script in the filesystem. However, ndm contains a  'generate-script' command in it's package.json. Given this I assumed the intent of this line is to run `ndm generate`. Running `ndm generate` successfully updated the daemon configuration changes in service.json.